requirements.txt
# Core dependencies for the weather pipeline project
requests==2.31.0
pandas==2.1.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.0

# Visualization (used in Jupyter Notebook)
matplotlib==3.8.2
notebook==7.0.7
ipykernel==6.29.0
